Crate pandoc

API that wraps the pandoc command line tool

Reexports

pub use DocumentClass::*;

Structs

Pandoc
the argument builder

Enums

DocumentClass
equivalent to the latex document class

EmailObfuscation

InputFormat
typesafe access to -f FORMAT, -r FORMAT, --from=FORMAT, --read=FORMAT

MarkdownExtension

OutputFormat
typesafe access to -t FORMAT, -w FORMAT, --to=FORMAT, --write=FORMAT

OutputFormatExt
allow to choose an output format with or without extensions

PandocError
Possible errors that can occur before or during pandoc execution

PandocOption

TrackChanges

Functions

new
does nothing useful, simply gives you a builder object convenience function so you can call pandoc::new()

Type Definitions

URL