Enum pandoc::PandocOutput

pub enum PandocOutput {
    ToFile(PathBuf),
    ToBuffer(String),
}

The output from Pandoc: the file written to, or a buffer with its output.

Variants
ToFile(PathBuf)

The results of the pandoc operation are stored in Path

ToBuffer(String)

The results of the pandoc operation are returned as a String