Struct pandoc::Pandoc

pub struct Pandoc { /* fields omitted */ }

the argument builder

Methods

impl Pandoc

fn new() -> Pandoc

Get a new Pandoc object This function returns a builder object to configure the Pandoc execution.

fn add_latex_path_hint<'p, T: AsRef<Path> + ?Sized>(
&'p mut self,
path: &T
) -> &'p mut Pandoc

Add a path hint to search for the LaTeX executable.

The supplied path is searched first for the latex executable, then the environment variable PATH, then some hard-coded location hints.

fn add_pandoc_path_hint<'p, T: AsRef<Path> + ?Sized>(
&'p mut self,
path: &T
) -> &'p mut Pandoc

Add a path hint to search for the Pandoc executable.

The supplied path is searched first for the Pandoc executable, then the environment variable PATH, then some hard-coded location hints.

fn set_doc_class<'p>(&'p mut self, class: DocumentClass) -> &'p mut Pandoc

Set or overwrite the document-class.

fn set_show_cmdline<'p>(&'p mut self, flag: bool) -> &'p mut Pandoc

Set whether Pandoc should print the used command-line

If set to true, the command-line to execute pandoc (as a subprocess) will be displayed on stdout.

fn set_output_format<'p>(
&'p mut self,
format: OutputFormat,
extensions: Vec<MarkdownExtension>
) -> &'p mut Pandoc

Set or overwrite the output format.

fn set_input_format<'p>(
&'p mut self,
format: InputFormat,
extensions: Vec<MarkdownExtension>
) -> &'p mut Pandoc

Set or overwrite the input format

fn add_input<'p, T: AsRef<Path> + ?Sized>(
&'p mut self,
filename: &T
) -> &'p mut Pandoc

Add additional input files

The order of adding the files is the order in which they are processed, hence the order is important. This function does not work, if input has been already set to standard input using set_input.

fn set_input(&mut self, input: InputKind) -> &mut Pandoc

Set input for Pandoc.

The input is given with pandoc::InputKind and overrides any inputs already supplied.

Example

fn main() {
    // pass in a string using standard input:
    let markdown = "**very** _important".into();
    let mut p = pandoc::new(); // assign to variable to increase life time
    p.set_input(pandoc::InputKind::Pipe(markdown));
}

fn set_output<'p>(&'p mut self, output: OutputKind) -> &'p mut Pandoc

Set or overwrite the output filename.

fn set_bibliography<'p, T: AsRef<Path> + ?Sized>(
&'p mut self,
filename: &T
) -> &'p mut Pandoc

Set the file name of the bibliography database.

fn set_csl<'p, T: AsRef<Path> + ?Sized>(
&'p mut self,
filename: &T
) -> &'p mut Pandoc

Set the filename of the citation style file.

fn set_toc<'p>(&'p mut self) -> &'p mut Pandoc

Enable the generation of a table of contents

By default, documents are transformed as they are. If this option is set, a table of contents is added right in front of the actual document.

fn set_chapters<'p>(&'p mut self) -> &'p mut Pandoc

Treat top-level headers as chapters in LaTeX, ConTeXt, and DocBook output.

fn set_number_sections<'p>(&'p mut self) -> &'p mut Pandoc

Set custom prefix for sections.

If this function is called, all sections will be numbered. Normally, sections in LaTeX, ConTeXt, HTML, or EPUB output are unnumbered.

fn set_latex_template<'p, T: AsRef<Path> + ?Sized>(
&'p mut self,
filename: &T
) -> &'p mut Pandoc

Set a custom latex template.

fn set_slide_level<'p>(&'p mut self, level: u32) -> &'p mut Pandoc

Set the header level that causes a new slide to be generated.

fn set_variable<'p, T: AsRef<str> + ?Sized, U: AsRef<str> + ?Sized>(
&'p mut self,
key: &T,
value: &U
) -> &'p mut Pandoc

Set a custom variable.

This method sets a custom Pandoc variable. It is adviced not to use this function, because there are convenience functions for most of the available variables.

fn add_filter<'p>(
&'p mut self,
filter: fn(_: String) -> String
) -> &'p mut Pandoc

Add a Pandoc filter.

Pandoc parses any of the supported input formats to an abstract syntax tree (AST). If a filter is specified, it will receive a JSON representation of this AST and can transform it to its liking and add/modify/remove elements. The output is then passed back to Pandoc.

fn add_option<'p>(&'p mut self, option: PandocOption) -> &'p mut Pandoc

Add a PandocOption.

fn add_options<'p>(&'p mut self, options: &[PandocOption]) -> &'p mut Pandoc

fn generate_latex_template<T: AsRef<str> + ?Sized>(self, filename: &T)

generate a latex template from the given settings

Warning: this function can panic in a lot of places.

fn execute(self) -> Result<PandocOutput, PandocError>

Execute the Pandoc configured command.

A successful Pandoc run can return either the path to a file written by the operation, or the result of the operation from stdio.

The PandocOutput variant returned depends on the OutputKind configured: