Crate pandoc

API that wraps the pandoc command line tool

Re-exports

pub use DocumentClass::*;

Structs

Pandoc
the argument builder

Enums

DocumentClass
equivalent to the latex document class

EmailObfuscation

InputFormat
typesafe access to -f FORMAT, -r FORMAT, --from=FORMAT, --read=FORMAT

InputKind

MarkdownExtension

OutputFormat
typesafe access to -t FORMAT, -w FORMAT, --to=FORMAT, --write=FORMAT

OutputKind
Specify whether to generate a file or pipe the output to stdout.

PandocError
Possible errors that can occur before or during pandoc execution

PandocOption

PandocOutput
The output from Pandoc: the file written to, or a buffer with its output.

Tld

TrackChanges

Functions

new
Convenience function to call Pandoc::new()

Type Definitions

URL