Struct pandoc::Pandoc

pub struct Pandoc {
    // some fields omitted
}

the argument builder

Methods

impl Pandoc

fn new() -> Pandoc

does nothing useful, simply gives you a builder object

fn add_latex_path_hint<T: AsRef<Path> + ?Sized>(&mut self, path: &T)

this path is searched first for latex, then PATH, then some hardcoded hints

fn add_pandoc_path_hint<T: AsRef<Path> + ?Sized>(&mut self, path: &T)

this path is searched first for pandoc, then PATH, then some hardcoded hints

fn set_doc_class(&mut self, class: DocumentClass)

sets or overwrites the document-class

fn set_output_format(&mut self, format: OutputFormat)

sets or overwrites the output format

fn set_input_format(&mut self, format: InputFormat)

sets or overwrites the input format

fn add_input<T: AsRef<Path> + ?Sized>(&mut self, filename: &T)

adds more input files, the order is relevant the order of adding the files is the order in which they are processed

fn set_output<T: AsRef<str> + ?Sized>(&mut self, filename: &T)

sets or overwrites the output filename

fn set_bibliography<T: AsRef<Path> + ?Sized>(&mut self, filename: &T)

filename of the bibliography database

fn set_csl<T: AsRef<Path> + ?Sized>(&mut self, filename: &T)

filename of a citation style file

fn set_toc(&mut self)

enable table of contents

fn set_chapters(&mut self)

enable chapters

fn set_number_sections(&mut self)

prefix section names with indices x.y.z

fn set_latex_template<T: AsRef<Path> + ?Sized>(&mut self, filename: &T)

set a custom latex template

fn set_slide_level(&mut self, level: u32)

sets the header level that causes a new slide to be generated

fn set_variable<T: AsRef<str> + ?Sized, U: AsRef<str> + ?Sized>(&mut self, key: &T, value: &U)

set a custom variable try not to use this, there are convenience functions for most things

fn add_filter(&mut self, filter: fn(String) -> String)

closures that take a json string and return a json string

fn add_option(&mut self, option: PandocOption)

fn generate_latex_template<T: AsRef<str> + ?Sized>(self, filename: &T)

generate a latex template from the given settings this function can panic in a lot of places

fn execute(self) -> Result<(), PandocError>

actually execute pandoc